// frontend/src/types/navigation.ts
export type AppPage =
  | "employees"
  | "dashboard"
  | "notifications"
  | "reports"
  | "profile"
  | "preferences"
  | "settings"
  | "admins"
  | "accessLogs"
  | "sendNote"
  | "leaveRequests"
  | "profileEdit"
  | "employeeLogins"
  | "messages"
  | "review-requests"
  | "threads"
  | "userDashboard"
  | "bulkActions"
  | "auditLogs"
  | "advancedEmployeeSearch"
  | "notificationInbox"
  | "messagingInbox"
  | "analyticsDashboard"
  | "employeeSelfServicePortal"
  | "slackIntegration";